Move the scroll-lock IPC focus tests into their per-behavior file
docs/TESTING.md puts new tests in files named for the behavior under
test, not
for the handler that happens to contain the code. Both scroll-lock
reveal tests
were sitting in IPCCommandRouterTests.swift, which is named for the
router.

They move verbatim into ScrollLockFocusCommandRevealTests.swift as a
second
suite, so the hotkey path and the IPC path — the one a workspace-bar
click and
nehirctl take — sit next to each other. The new suite needs @MainActor
for the
same reason the router suite has it.

Three helpers the tests rely on lose their `private` scope so the moved
suite
can reach them: ipcCommandRouterSessionToken, makeIPCCommandRouter and
prepareIPCNiriState. Copying them instead would leave two definitions to
drift
apart.
